lgc: Separate BuilderCommon and BuilderBase
BuilderBase is where we put methods for generating IR that are used in
both BuilderImpl subclasses (e.g. InOutBuilder) and in LGC passes.

The problem was that the class hierarchy (where Builder inherits from
BuilderBase and BuilderImplBase inherits from Builder) means that those
methods are visible to front-end code outside LGC. Those methods are not
intended to be used from there.

This commit moves BuilderBase out of the LGC interface and makes it no
longer the superclass of Builder. There is a new BuilderCommon class
occupying BuilderBase's former position in the hierarchy, containing
only CreateNamedCall(), which is legitimately used by both the LLPC
front-end and LGC.

To allow BuilderImpl subclasses such as InOutBuilder to access the
BuilderBase methods, I added a static method BuilderBase::get to static
cast a BuilderCommon into BuilderBase, and some forwarding methods in
BuilderImplBase for BuilderBase's methods.

Also I moved CreateMapToInt32 out of Builder and in to BuilderBase.
