+++ lgc/builder/BuilderImpl.h
@@ -32,6 +32,7 @@
 
 #include "lgc/Builder.h"
 #include "lgc/state/PipelineState.h"
+#include "lgc/util/BuilderBase.h"
 
 namespace lgc {
 
@@ -92,6 +93,21 @@ protected:
   // Create code to get the lane number within the wave. This depends on whether the shader is wave32 or wave64,
   // and thus on the shader stage it is used from.
   llvm::Value *CreateGetLaneNumber();
+
+  // Forwarding methods for methods in BuilderBase that BuilderImpl subclasses may need to access.
+  // We want these methods in BuilderBase to be accessible from anywhere in LGC, both BuilderImpl subclasses
+  // and later passes, but not from outside LGC. There is no possible class hierarchy to make that happen, without
+  // also making Builder methods visible from later passes in LGC, which we don't want.
+  llvm::Value *CreateRelocationConstant(const llvm::Twine &symbolName) {
+    return BuilderBase::get(*this).CreateRelocationConstant(symbolName);
+  }
+  llvm::Value *CreateAddByteOffset(llvm::Value *pointer, llvm::Value *byteOffset, const llvm::Twine &instName = "") {
+    return BuilderBase::get(*this).CreateAddByteOffset(pointer, byteOffset, instName);
+  }
+  llvm::Value *CreateMapToInt32(BuilderBase::MapToInt32Func mapFunc, llvm::ArrayRef<llvm::Value *> mappedArgs,
+                                llvm::ArrayRef<llvm::Value *> passthroughArgs) {
+    return BuilderBase::get(*this).CreateMapToInt32(mapFunc, mappedArgs, passthroughArgs);
+  }
 
   PipelineState *m_pipelineState = nullptr; // Pipeline state
