+++ lgc/patch/PatchWorkarounds.cpp
@@ -30,10 +30,10 @@
  */
 
 #include "PatchWorkarounds.h"
-#include "lgc/BuilderBase.h"
 #include "lgc/state/PipelineShaders.h"
 #include "lgc/state/PipelineState.h"
 #include "lgc/state/TargetInfo.h"
+#include "lgc/util/BuilderBase.h"
 #include "llvm/IR/Constants.h"
 #include "llvm/IR/Instructions.h"
 #include "llvm/IR/IntrinsicsAMDGPU.h"
