+++ lgc/patch/VertexFetch.cpp
@@ -29,7 +29,6 @@
  ***********************************************************************************************************************
  */
 #include "lgc/patch/VertexFetch.h"
-#include "lgc/BuilderBase.h"
 #include "lgc/LgcContext.h"
 #include "lgc/patch/Patch.h"
 #include "lgc/patch/ShaderInputs.h"
@@ -37,6 +36,7 @@
 #include "lgc/state/PalMetadata.h"
 #include "lgc/state/PipelineState.h"
 #include "lgc/state/TargetInfo.h"
+#include "lgc/util/BuilderBase.h"
 #include "lgc/util/Internal.h"
 #include "llvm/IR/Constants.h"
 #include "llvm/IR/Instructions.h"
